requirements.txt
selenium==4.14.0
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
openpyxl==3.11.0
python-docx==0.8.11
docxtpl==0.16.8
webdriver-manager==4.0.1
requests==2.31.0
